+++ packages/generator/src/openapi/index.ts
@@ -3,6 +3,7 @@ import FileGenerator from "./fileGenerator";
 import ModelProcessor from "./modelProcessor";
 import { IConfig, IGeneratorParams } from "./types";
 import NameFormatter from "./formatters/nameFormatter";
+import ObservableFormatter from "./formatters/observableFormatter";
 
 export default class OpenApiGenerator extends GeneratorBase<IGeneratorParams, IConfig> {
   async run() {
@@ -14,9 +15,11 @@ export default class OpenApiGenerator extends GeneratorBase<IGeneratorParams, IC
     const modelProcessor = new ModelProcessor();
     const { entities, enums } = await modelProcessor.process(this.config.api);
 
-    const formatter = new NameFormatter();
-    formatter.formatEnums(enums);
-    formatter.formatEntities(entities);
+    const nameFormatter = new NameFormatter();
+    nameFormatter.formatEnums(enums);
+    nameFormatter.formatEntities(entities);
+
+    new ObservableFormatter().formatEntities(this.config.observable, entities);
 
     const fileGenerator = new FileGenerator(this.project, entities, enums);
     await fileGenerator.generate(this.params);